pyproject.toml
[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry]
name = "gliff"
version = "0.1.0"
description = "gliff.ai SDK - a Python SDL for developing AI-in-the-loop trusted services for the gliff.ai platform"
authors = ["Chas Nelson <chas@gliff.ai>"]
license = "AGPL-3.0-or-later"
include = ["LICENSE",]

[tool.poetry.dependencies]
python = "^3.9"
python-decouple = "^3.5"
loguru = "^0.5.3"
etebase = "^0.31.2"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
black = "^21.10b0"

[tool.poetry.scripts]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
